# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: SimulationForge API
  description: Simulations, surfaces, incertitude, convergence, dérive et couverture
    explicite de scénarios déterministes, reproductibles et non prédictives.
  version: 1.0.7
paths:
...
  /v1/models:
    get:
      summary: List Models
      operationId: list_models_v1_models_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/SimulationModel'
                type: array
                title: Response List Models V1 Models Get
    post:
      summary: Create Model
      operationId: create_model_v1_models_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimulationModelCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulationModel'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/models/{model_id}:
    get:
      summary: Get Model
      operationId: get_model_v1_models__model_id__get
      parameters:
      - name: model_id
        in: path
        required: true
        schema:
          type: string
          title: Model Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulationModel'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/scenarios:
    get:
      summary: List Scenarios
      operationId: list_scenarios_v1_scenarios_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Scenario'
                type: array
                title: Response List Scenarios V1 Scenarios Get
    post:
      summary: Create Scenario
      operationId: create_scenario_v1_scenarios_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScenarioCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scenario'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/scenarios/{scenario_id}:
    get:
      summary: Get Scenario
      operationId: get_scenario_v1_scenarios__scenario_id__get
      parameters:
      - name: scenario_id
        in: path
        required: true
        schema:
          type: string
          title: Scenario Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scenario'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/runs:
    get:
      summary: List Runs
      operationId: list_runs_v1_runs_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/SimulationRun'
                type: array
                title: Response List Runs V1 Runs Get
    post:
      summary: Run Simulation
      operationId: run_simulation_v1_runs_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimulationRunCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulationRun'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/runs/{run_id}:
    get:
      summary: Get Run
      operationId: get_run_v1_runs__run_id__get
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulationRun'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/comparisons:
    get:
      summary: List Comparisons
      operationId: list_comparisons_v1_comparisons_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ScenarioComparison'
                type: array
                title: Response List Comparisons V1 Comparisons Get
    post:
      summary: Compare Scenarios
      operationId: compare_scenarios_v1_comparisons_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScenarioComparisonCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScenarioComparison'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/comparisons/{comparison_id}:
    get:
      summary: Get Comparison
      operationId: get_comparison_v1_comparisons__comparison_id__get
      parameters:
      - name: comparison_id
        in: path
        required: true
        schema:
          type: string
          title: Comparison Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScenarioComparison'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/sensitivities:
    get:
      summary: List Sensitivities
      operationId: list_sensitivities_v1_sensitivities_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/SensitivityAnalysis'
                type: array
                title: Response List Sensitivities V1 Sensitivities Get
    post:
      summary: Analyze Sensitivity
      operationId: analyze_sensitivity_v1_sensitivities_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SensitivityAnalysisCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SensitivityAnalysis'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/sensitivities/{analysis_id}:
    get:
      summary: Get Sensitivity
      operationId: get_sensitivity_v1_sensitivities__analysis_id__get
      parameters:
      - name: analysis_id
        in: path
        required: true
        schema:
          type: string
          title: Analysis Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SensitivityAnalysis'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/interaction-surfaces:
    get:
      summary: List Interaction Surfaces
      operationId: list_interaction_surfaces_v1_interaction_surfaces_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/InteractionSurface'
                type: array
                title: Response List Interaction Surfaces V1 Interaction Surfaces
                  Get
    post:
      summary: Analyze Interaction Surface
      operationId: analyze_interaction_surface_v1_interaction_surfaces_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InteractionSurfaceCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InteractionSurface'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/interaction-surfaces/{surface_id}:
    get:
      summary: Get Interaction Surface
      operationId: get_interaction_surface_v1_interaction_surfaces__surface_id__get
      parameters:
      - name: surface_id
        in: path
        required: true
        schema:
          type: string
          title: Surface Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InteractionSurface'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/uncertainty-dossiers:
    get:
      summary: List Uncertainty Dossiers
      operationId: list_uncertainty_dossiers_v1_uncertainty_dossiers_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/UncertaintyDossier'
                type: array
                title: Response List Uncertainty Dossiers V1 Uncertainty Dossiers
                  Get
    post:
      summary: Create Uncertainty Dossier
      operationId: create_uncertainty_dossier_v1_uncertainty_dossiers_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UncertaintyDossierCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UncertaintyDossier'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/uncertainty-dossiers/{dossier_id}:
    get:
      summary: Get Uncertainty Dossier
      operationId: get_uncertainty_dossier_v1_uncertainty_dossiers__dossier_id__get
      parameters:
      - name: dossier_id
        in: path
        required: true
        schema:
          type: string
          title: Dossier Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UncertaintyDossier'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/convergence-dossiers:
    get:
      summary: List Convergence Dossiers
      operationId: list_convergence_dossiers_v1_convergence_dossiers_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ConvergenceDossier'
                type: array
                title: Response List Convergence Dossiers V1 Convergence Dossiers
                  Get
    post:
      summary: Create Convergence Dossier
      operationId: create_convergence_dossier_v1_convergence_dossiers_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConvergenceDossierCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConvergenceDossier'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/convergence-dossiers/{dossier_id}:
    get:
      summary: Get Convergence Dossier
      operationId: get_convergence_dossier_v1_convergence_dossiers__dossier_id__get
      parameters:
      - name: dossier_id
        in: path
        required: true
        schema:
          type: string
          title: Dossier Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConvergenceDossier'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/scenario-drift-dossiers:
    get:
      summary: List Scenario Drift Dossiers
      operationId: list_scenario_drift_dossiers_v1_scenario_drift_dossiers_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ScenarioDriftDossier'
                type: array
                title: Response List Scenario Drift Dossiers V1 Scenario Drift Dossiers
                  Get
    post:
      summary: Create Scenario Drift Dossier
      operationId: create_scenario_drift_dossier_v1_scenario_drift_dossiers_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScenarioDriftDossierCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScenarioDriftDossier'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/scenario-drift-dossiers/{dossier_id}:
    get:
      summary: Get Scenario Drift Dossier
      operationId: get_scenario_drift_dossier_v1_scenario_drift_dossiers__dossier_id__get
      parameters:
      - name: dossier_id
        in: path
        required: true
        schema:
          type: string
          title: Dossier Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScenarioDriftDossier'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/scenario-coverage-dossiers:
    get:
      summary: List Scenario Coverage Dossiers
      operationId: list_scenario_coverage_dossiers_v1_scenario_coverage_dossiers_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ScenarioCoverageDossier'
                type: array
                title: Response List Scenario Coverage Dossiers V1 Scenario Coverage
                  Dossiers Get
    post:
      summary: Create Scenario Coverage Dossier
      operationId: create_scenario_coverage_dossier_v1_scenario_coverage_dossiers_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScenarioCoverageDossierCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScenarioCoverageDossier'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/scenario-coverage-dossiers/{dossier_id}:
    get:
      summary: Get Scenario Coverage Dossier
      operationId: get_scenario_coverage_dossier_v1_scenario_coverage_dossiers__dossier_id__get
      parameters:
      - name: dossier_id
        in: path
        required: true
        schema:
          type: string
          title: Dossier Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScenarioCoverageDossier'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ConvergenceDossierCreate:
      properties:
        run_ids:
          items:
            type: string
          type: array
          maxItems: 50
          minItems: 3
          title: Run Ids
      additionalProperties: false
      type: object
      required:
      - run_ids
      title: ConvergenceDossierCreate
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Scenario:
      properties:
        id:
          type: string
          title: Id
        scenario_hash:
          type: string
          title: Scenario Hash
        model_hash:
          type: string
          title: Model Hash
        specification:
          $ref: '#/components/schemas/ScenarioCreate'
        created_at:
          type: string
          format: date-time
          title: Created At
      additionalProperties: false
      type: object
      required:
      - id
      - scenario_hash
      - model_hash
      - specification
      - created_at
      title: Scenario
...
    ScenarioComparisonCreate:
      properties:
        baseline_scenario_id:
          type: string
          maxLength: 64
          minLength: 1
          title: Baseline Scenario Id
        stress_scenario_id:
          type: string
          maxLength: 64
          minLength: 1
          title: Stress Scenario Id
        seed:
          type: integer
          maximum: 9.223372036854776e+18
          minimum: -9.223372036854776e+18
          title: Seed
        iterations:
          type: integer
          maximum: 10000.0
          minimum: 1.0
          title: Iterations
      additionalProperties: false
      type: object
      required:
      - baseline_scenario_id
      - stress_scenario_id
      - seed
      - iterations
      title: ScenarioComparisonCreate
...
    ScenarioCreate:
      properties:
        model_id:
          type: string
          maxLength: 64
          minLength: 1
          title: Model Id
        name:
          type: string
          pattern: ^[A-Za-z][A-Za-z0-9_-]{0,63}$
          title: Name
        version:
          type: string
          pattern: ^[0-9]+\.[0-9]+\.[0-9]+$
          title: Version
        description:
          type: string
          maxLength: 500
          minLength: 3
          title: Description
        parameter_overrides:
          additionalProperties:
            type: number
            maximum: 1000000.0
            minimum: -1000000.0
          type: object
          maxProperties: 32
          title: Parameter Overrides
        assumptions:
          items:
            type: string
          type: array
          maxItems: 32
          title: Assumptions
          default: []
      additionalProperties: false
      type: object
      required:
      - model_id
      - name
      - version
      - description
      title: ScenarioCreate
...
    ScenarioDriftDossierCreate:
      properties:
        run_ids:
          items:
            type: string
          type: array
          maxItems: 100
          minItems: 2
          title: Run Ids
      additionalProperties: false
      type: object
      required:
      - run_ids
      title: ScenarioDriftDossierCreate
...
    SimulationModel:
      properties:
        id:
          type: string
          title: Id
        model_hash:
          type: string
          title: Model Hash
        specification:
          $ref: '#/components/schemas/SimulationModelCreate'
        created_at:
          type: string
          format: date-time
          title: Created At
      additionalProperties: false
      type: object
      required:
      - id
      - model_hash
      - specification
      - created_at
      title: SimulationModel
    SimulationModelCreate:
      properties:
        name:
          type: string
          pattern: ^[A-Za-z][A-Za-z0-9_-]{0,63}$
          title: Name
        version:
          type: string
          pattern: ^[0-9]+\.[0-9]+\.[0-9]+$
          title: Version
        summary:
          type: string
          maxLength: 500
          minLength: 3
          title: Summary
        assumptions:
          items:
            type: string
          type: array
          maxItems: 32
          minItems: 1
          title: Assumptions
        variables:
          items:
            $ref: '#/components/schemas/VariableSpec'
          type: array
          maxItems: 32
          minItems: 1
          title: Variables
        outcome:
          $ref: '#/components/schemas/OutcomeSpec'
      additionalProperties: false
      type: object
      required:
      - name
      - version
      - summary
      - assumptions
      - variables
      - outcome
      title: SimulationModelCreate
    SimulationRun:
      properties:
        id:
          type: string
          title: Id
        scenario_id:
          type: string
          title: Scenario Id
        scenario_hash:
          type: string
          title: Scenario Hash
        seed:
          type: integer
          title: Seed
        iterations:
          type: integer
          title: Iterations
        algorithm_version:
          type: string
          title: Algorithm Version
        statistics:
          $ref: '#/components/schemas/SimulationStatistics'
        qualification:
          type: string
          title: Qualification
        warning:
          type: string
          title: Warning
        result_hash:
          type: string
          title: Result Hash
        reproducible:
          type: boolean
          title: Reproducible
        idempotent_replay:
          type: boolean
          title: Idempotent Replay
        created_at:
          type: string
          format: date-time
          title: Created At
      additionalProperties: false
      type: object
      required:
      - id
      - scenario_id
      - scenario_hash
      - seed
      - iterations
      - algorithm_version
      - statistics
      - qualification
      - warning
      - result_hash
      - reproducible
      - idempotent_replay
      - created_at
      title: SimulationRun
    SimulationRunCreate:
      properties:
        scenario_id:
          type: string
          maxLength: 64
          minLength: 1
          title: Scenario Id
        seed:
          type: integer
          maximum: 9.223372036854776e+18
          minimum: -9.223372036854776e+18
          title: Seed
        iterations:
          type: integer
          maximum: 10000.0
          minimum: 1.0
          title: Iterations
      additionalProperties: false
      type: object
      required:
      - scenario_id
      - seed
      - iterations
      title: SimulationRunCreate
...
    UncertaintyDossierCreate:
      properties:
        run_ids:
          items:
            type: string
          type: array
          maxItems: 100
          minItems: 2
          title: Run Ids
        limits:
          items:
            $ref: '#/components/schemas/MetricLimit'
          type: array
          maxItems: 7
          title: Limits
          default: []
      additionalProperties: false
      type: object
      required:
      - run_ids
      title: UncertaintyDossierCreate
...